+++ Library/Compiler.php
@@ -466,6 +466,7 @@ final class Compiler
      * Returns a Zephir Constant by its name.
      *
      * @param string $name
+     *
      * @return mixed
      */
     public function getConstant(string $name)
@@ -853,7 +854,7 @@ final class Compiler
     /**
      * Compiles the extension without installing it.
      *
-     * @param bool $development
+     * @param bool     $development
      * @param int|null $jobs
      *
      * @throws Exception
@@ -1324,7 +1325,7 @@ final class Compiler
                     $structBuilder->addProperty($field, $global);
 
                     $isModuleGlobal = (int) !empty($global['module']);
-                    $globalsDefault[$isModuleGlobal][] = $structBuilder->getCDefault($field, $global, $namespace).PHP_EOL;
+                    $globalsDefault[$isModuleGlobal][] = $structBuilder->getCDefault($field, $global, $namespace);
                     $initEntries[] = $structBuilder->getInitEntry($field, $global, $namespace);
                 }
 
@@ -1333,8 +1334,7 @@ final class Compiler
 
             $globalCode = PHP_EOL;
             foreach ($structures as $structureName => $internalStructure) {
-                $globalCode .= "\t".'zephir_struct_'.$structureName.' '.
-                                $structureName.';'.PHP_EOL.PHP_EOL;
+                $globalCode .= "\t".'zephir_struct_'.$structureName.' '.$structureName.';'.PHP_EOL;
             }
 
             /**
@@ -1351,16 +1351,16 @@ final class Compiler
 
                 $isModuleGlobal = (int) !empty($global['module']);
                 $type = $global['type'];
-                // TODO: Add support for 'string', 'hash'
+                // TODO: Add support for 'hash'
                 // TODO: Zephir\Optimizers\FunctionCall\GlobalsSetOptimizer
                 switch ($global['type']) {
                     case 'boolean':
                     case 'bool':
                         $type = 'zend_bool';
                         if (true === $global['default']) {
-                            $globalsDefault[$isModuleGlobal][] = "\t".$namespace.'_globals->'.$name.' = 1;'.PHP_EOL;
+                            $globalsDefault[$isModuleGlobal][] = "\t".$namespace.'_globals->'.$name.' = 1;';
                         } else {
-                            $globalsDefault[$isModuleGlobal][] = "\t".$namespace.'_globals->'.$name.' = 0;'.PHP_EOL;
+                            $globalsDefault[$isModuleGlobal][] = "\t".$namespace.'_globals->'.$name.' = 0;';
                         }
                         break;
 
@@ -1368,16 +1368,16 @@ final class Compiler
                     case 'uint':
                     case 'long':
                     case 'double':
-                        $globalsDefault[$isModuleGlobal][]
-                            = "\t".$namespace.'_globals->'.$name.' = '.
-                            $global['default'].';'.PHP_EOL;
+                        $globalsDefault[$isModuleGlobal][] = "\t".$namespace.'_globals->'.$name.' = '.$global['default'].';';
                         break;
 
                     case 'char':
                     case 'uchar':
-                        $globalsDefault[$isModuleGlobal][]
-                            = "\t".$namespace.'_globals->'.$name.' = \''.
-                            $global['default'].'\';'.PHP_EOL;
+                        $globalsDefault[$isModuleGlobal][] = "\t".$namespace.'_globals->'.$name.' = \''.$global['default'].'\';';
+                        break;
+                    case 'string':
+                        $type = 'char *';
+                        $globalsDefault[$isModuleGlobal][] = "\t".$namespace.'_globals->'.$name.' = ZSTR_VAL(zend_string_init(ZEND_STRL("'.$global['default'].'"), 0));';
                         break;
                     default:
                         throw new Exception(
@@ -1385,7 +1385,7 @@ final class Compiler
                         );
                 }
 
-                $globalCode .= "\t".$type.' '.$name.';'.PHP_EOL.PHP_EOL;
+                $globalCode .= "\t".$type.' '.$name.';'.PHP_EOL;
                 $iniEntry = [];
                 if (isset($global['ini-entry'])) {
                     $iniEntry = $global['ini-entry'];
@@ -1424,8 +1424,8 @@ final class Compiler
             }
         }
 
-        $globalsDefault[0] = implode('', $globalsDefault[0]);
-        $globalsDefault[1] = implode('', $globalsDefault[1]);
+        $globalsDefault[0] = implode(PHP_EOL, $globalsDefault[0]);
+        $globalsDefault[1] = implode(PHP_EOL, $globalsDefault[1]);
 
         return [$globalCode, $globalStruct, $globalsDefault, $initEntries];
     }
@@ -1452,7 +1452,7 @@ final class Compiler
                     $headerArray[] = '"'.htmlentities($header).'"';
                 }
 
-                $phpinfo .= "\t".'php_info_print_table_header('. count($headerArray).', '.
+                $phpinfo .= "\t".'php_info_print_table_header('.count($headerArray).', '.
                     implode(', ', $headerArray).');'.PHP_EOL;
             }
 
@@ -1463,7 +1463,7 @@ final class Compiler
                         $rowArray[] = '"'.htmlentities($field).'"';
                     }
 
-                    $phpinfo .= "\t".'php_info_print_table_row('. count($rowArray).', '.
+                    $phpinfo .= "\t".'php_info_print_table_row('.count($rowArray).', '.
                         implode(', ', $rowArray).');'.PHP_EOL;
                 }
             }
@@ -2191,7 +2191,7 @@ final class Compiler
     /**
      * Process config.w32 sections.
      *
-     * @param array $sources
+     * @param array  $sources
      * @param string $project
      *
      * @return array