+++ ext/stub.c
@@ -240,6 +240,7 @@ zend_class_entry *stub_trytest_ce;
 zend_class_entry *stub_typehinting_testabstract_ce;
 zend_class_entry *stub_typeinstances_ce;
 zend_class_entry *stub_typeoff_ce;
+zend_class_entry *stub_types_mixedtype_ce;
 zend_class_entry *stub_unknownclass_ce;
 zend_class_entry *stub_unsettest_ce;
 zend_class_entry *stub_usetest_ce;
@@ -465,6 +466,7 @@ static PHP_MINIT_FUNCTION(stub)
 	ZEPHIR_INIT(Stub_TypeHinting_TestAbstract);
 	ZEPHIR_INIT(Stub_TypeInstances);
 	ZEPHIR_INIT(Stub_Typeoff);
+	ZEPHIR_INIT(Stub_Types_MixedType);
 	ZEPHIR_INIT(Stub_UnknownClass);
 	ZEPHIR_INIT(Stub_Unsettest);
 	ZEPHIR_INIT(Stub_UseTest);
@@ -524,7 +526,7 @@ static void php_zephir_init_globals(zend_stub_globals *stub_globals)
 	stub_globals->my_setting_2 = 10;
 	stub_globals->my_setting_3 = 15.2;
 	stub_globals->my_setting_4 = 'A';
-
+	stub_globals->my_setting_5 = ZSTR_VAL(zend_string_init(ZEND_STRL("custom_value"), 0));
 	
 }