+++ tests/fixtures/stubs/issues/expected/Issue_1907.php
@@ -16,5 +16,4 @@ class Issue_1907 extends Exception implements \Throwable
     public function __construct()
     {
     }
-
 }